+++ desktop-widgets/tab-widgets/maintab.cpp
@@ -79,7 +79,7 @@ MainTab::MainTab(QWidget *parent) : QTabWidget(parent),
 	ui.weights->setModel(weightModel);
 	closeMessage();
 
-	connect(&diveListNotifier, &DiveListNotifier::divesEdited, this, &MainTab::divesEdited);
+	connect(&diveListNotifier, &DiveListNotifier::divesChanged, this, &MainTab::divesChanged);
 
 	connect(ui.editDiveSiteButton, SIGNAL(clicked()), MainWindow::instance(), SIGNAL(startDiveSiteEdit()));
 	connect(ui.location, &DiveLocationLineEdit::entered, MapWidget::instance(), &MapWidget::centerOnIndex);
@@ -338,10 +338,10 @@ static void profileFromDive(struct dive *d)
 
 // This function gets called if a field gets updated by an undo command.
 // Refresh the corresponding UI field.
-void MainTab::divesEdited(const QVector<dive *> &, DiveField field)
+void MainTab::divesChanged(dive_trip *trip, const QVector<dive *> &dives, DiveField field)
 {
-	// If there is no current dive, no point in updating anything
-	if (!current_dive)
+	// If the current dive is not in list of changed dives, do nothing
+	if (!current_dive || !dives.contains(current_dive))
 		return;
 
 	switch(field) {
